Оптимизировал структуру Добавил requirements.txt Дополнил readme Объединил все в одну базу
