+++ dev_notebooks/chamorcev_notebook.md
@@ -121,4 +121,21 @@ https://t.me/NetologyStudent_bot
 то не стал переносить все в общую базу, а прикрутил отдельно стоящую базу с вопросами.
 2. Нужно провести рефакторинг кода и повыносить методы в отдельные файлы.
 В общем заезжаем с тех. долгом.
- 
+
+2024-05-08 
+Начал работы по рефакторингу базы. 
+Объединил две базы в одну. Добавил подключение к postgree пока оставил sqllite но при запуске проекта 
+в бой можно мигрировать.
+
+2024.05.14
+Причесал код, сделал рефакторинг в админке
+Перене конфиг и все что с ним связано в папку конфиг
+Все что связано с базой положил в папку database
+nlp - тут все что по работе с поисковым методом ответов.
+
+2024.05.15
+Нужно бы Readme привести в порядок. Постараемся заняться на днях.
+
+2024.05.21 
+Добавил requements.txt
+Добавил чуть чуть описания в Readme
