+++ requirements.txt
@@ -1,0 +1,7 @@
+aiogram==3.6.0
+Flask==3.0.3
+nltk==3.8.1
+pymorphy3==2.0.1
+PyYAML==6.0.1
+PyYAML==6.0.1
+SQLAlchemy==2.0.27
