+++ src/admin/templates/add_question.html
@@ -12,6 +12,10 @@
           <label>Ответ</label>
           <textarea name='answer' cols="40" rows="5" required class='form-control'></textarea>
         </div>
+        <div class='form-group'>
+          <label>Ключевые слова</label>
+          <input type='text' name='key_words' required class='form-control'>
+        </div>
         <input type='submit' value='Submit' class='btn btn-primary'>
       </form>
