+++ src/admin/templates/edit_question.html
@@ -12,6 +12,10 @@
           <label>Ответ</label>
           <textarea name='answer' cols="40" rows="10" required class='form-control' >{{datas.answer}}</textarea>
         </div>
+        <div class='form-group'>
+          <label>Ключевые слова</label>
+          <input type='text' name='key_words' required class='form-control' value='{{datas.key_words}}'>
+        </div>
         <input type='submit' value='Submit' class='btn btn-primary'>
       </form>
     </div>
