+++ src/admin/templates/index.html
@@ -1,20 +1,22 @@
 {% extends 'layout.html' %}
 
 {% block body %}
-  <h3 class='text-center text-muted mb-3'>Админ панель для модерирования вопросов</h3>
-<table align="right">
-  <th>
-  <td ><p class='text-right'><a href='{{url_for("menu")}}' class='btn btn-light '>Menu</a></p></td>
-  <td ><p class='text-right'><a href='{{url_for("logout")}}' class='btn btn-light '>Logout</a></p></td>
-  </th>
-</table>
-<p class='text-right'><a href='{{url_for("add_question")}}' class='btn btn-success '>+Добавить вопрос</a></p>
-  <table class='table table-bordered'>
+<div class="sticky-top text-center text-muted mb-3 bg-light">
+    <h3 class='text-center text-muted mb-3'>Админ панель для вопросов</h3>
+    <table align="right">
+      <th>
+        <td ><p class='text-right'><a href='{{url_for("menu")}}' class='btn btn-light '>Menu</a></p></td>
+        <td ><p class='text-right'><a href='{{url_for("logout")}}' class='btn btn-light '>Logout</a></p></td>
+      </th>
+    </table>
+    <p class='text-right'><a href='{{url_for("add_question")}}' class='btn btn-success '>+Добавить вопрос</a></p>
+</div>
+<table class='table table-bordered'>
     <thead>
       <th>ID</th>
       <th>Вопрос</th>
       <th>Ответ</th>
-      <th>Точность</th>
+      <th>Ключевые слова</th>
       <th>Редактировать</th>
       <th>Удалить</th>
     </thead>
@@ -24,6 +26,7 @@
           <td>{{loop.index}}</td>
           <td>{{row.question}}</td>
           <td>{{row.answer}}</td>
+          <td>{{row.key_words}}</td>
           <td><a href='{{url_for("edit_question",id=row.id)}}' class='btn btn-primary'>Редактировать</a></td>
           <td><a href='{{url_for("delete_question",id=row.id)}}' class='btn btn-danger' onclick='return confirm("Вы уверены?")'>Удалить</a></td>
         </tr>
