+++ src/database/db_model.py
@@ -1,12 +1,15 @@
 from sqlalchemy import create_engine, Column, Integer, Text, ForeignKey, String, Boolean, DateTime
-from config_loader import load_config
-from sqlalchemy.orm import relationship, declarative_base
+from src.config.config_loader import load_config
+from sqlalchemy.orm import declarative_base
 from datetime import datetime
+import sys
+import os
+sys.path.append(os.path.join(os.getcwd(), '../..'))
 
 # Загружаем данные из конфига
 
 
-config = load_config("config.yaml")
+config = load_config("../config/config.yaml")
 
 # Присваиваем переменные для дальнейшего более простого использования
 bot_token = config['Telegram']['token']
@@ -20,7 +23,10 @@ db_host = config['Database']['host']
 # Соединяемся с базой данных Postgresql
 # engine = create_engine(f'postgresql://{db_user}:{pwd}@{db_host}:5432/{db_name}')
 # Для соединения с sqlLite
-engine = create_engine('sqlite:///sqlite3.db')
+
+db = os.path.join('c:/Users/chamorcev/PycharmProjects/EduAssistant/src/database/', 'sqlite3.db')
+
+engine = create_engine(f'sqlite:///{db}')
 
 
 Base = declarative_base()
@@ -46,23 +52,19 @@ class Menu(Base):
     name = Column(String(20), nullable=False)
     url = Column(String(10), nullable=False)
     active = Column(Boolean, default=False)
+    answer = Column(Text, default=False)
     created_on = Column(DateTime, default=datetime.now)
     updated_on = Column(DateTime, default=datetime.now)
 
 
-class Answer(Base):
-    __tablename__ = 'answers'
+class Questions(Base):
+    __tablename__ = 'questions'
 
     id = Column(Integer, primary_key=True)
-    menu_id = Column(Integer, ForeignKey('menu.id'))
-    author_id = Column(Integer, ForeignKey('users.id'))
+    question = Column(Text, nullable=False)
     answer = Column(Text, nullable=False)
-    active = Column(Boolean, default=False)
-    created_on = Column(DateTime, default=datetime.now)
-    updated_on = Column(DateTime, default=datetime.now)
-
-    menu = relationship("Menu")
-    author = relationship("User")
+    score = Column(Text, nullable=False)
+    keywords = Column(Text, nullable=False)
 
 
 Base.metadata.create_all(engine)
