+++ src/edu_bot.py
@@ -4,11 +4,11 @@ from aiogram.enums import ParseMode
 from aiogram.filters import CommandStart, Command
 from aiogram.types import Message
 from aiogram.utils.markdown import hbold
-from sqlalchemy.orm import sessionmaker
 
-from db_model import Menu, Answer, User
+
+from src.database.db_model import Menu,  User
 from utils import singleton
-from main_menu import url_dict, session, set_main_menu
+from menu_commands import url_dict, session, set_main_menu
 from nlp.find_answer import find_answer
 
 @singleton
@@ -47,8 +47,7 @@ class EduBot:
             try:
                 # Проверяем есть ли значение в базе. Если есть то отдаем ответ
                 if message.text in url_dict:
-                    answer = session.query(Answer.answer).join(Menu, Menu.id == Answer.menu_id).filter(
-                        Menu.url == message.text).all()
+                    answer = session.query(Menu.answer).filter(Menu.url == message.text).all()
                     for row in answer:
                         await message.answer(f"{row.answer}")
                 else:
@@ -61,8 +60,8 @@ class EduBot:
                     # (в данный момент ссылается на самого пользователя)
                     else:
                         await message.answer(
-                            f'Похоже вашего вопроса нет в базе данных, но вы можете задать его своему '
-                            f'<b><a href="tg://user?id={message.from_user.id}">куратору</a></b>')
+                            f'Похоже я не знаю ответа на ваш вопрос, но вы можете задать его своему '
+                            f'<b><a href="https://t.me/kbaskakova_neto_vo">куратору</a></b>')
 
             except TypeError:
                 # Если что то пошло не так выводим эту строчку
